Fix favorite handling with new 'innards'.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@51907 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ elmdentica/src/twitter.h
@@ -89,8 +89,8 @@ typedef struct _user_get {
 
 int ed_twitter_post(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, char *msg);
 void ed_twitter_timeline_get(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, int timeline);
-void ed_twitter_favorite_create(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, long int status_id);
-void ed_twitter_favorite_destroy(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, long int status_id);
+void ed_twitter_favorite_create(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, long long int status_id);
+void ed_twitter_favorite_destroy(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, long long int status_id);
 void ed_twitter_init_friends(void);
 anUser *ed_twitter_user_get(int account_id, UserProfile *user);
 void ed_twitter_user_follow(int account_id, char *screen_name, char *password, char *proto, char *domain, int port, char *base_url, char *user_screen_name);
